Create KNN and Per Instance Algorithm Selection
KNN - requires dataset from LR and LGBM 
For each algorithm that was best within a cluster, the algorithm is trained on the training set and results gotten from test set. 
Then depending on which cluster the instance is assigned to, the best performing algorithm is chosen to predict. 
Here all predictions are calculated and then one is selected - this is not ideal. Should probably have a function to do automatically.